+++ internal/platform/browseropen/open.go
@@ -1,0 +1,29 @@
+package browseropen
+
+import (
+	"fmt"
+	"os/exec"
+	"runtime"
+)
+
+type Opener interface {
+	Open(url string) error
+}
+
+type DefaultOpener struct{}
+
+func (DefaultOpener) Open(url string) error {
+	if url == "" {
+		return fmt.Errorf("empty url")
+	}
+	var cmd *exec.Cmd
+	switch runtime.GOOS {
+	case "darwin":
+		cmd = exec.Command("open", url)
+	case "windows":
+		cmd = exec.Command("rundll32", "url.dll,FileProtocolHandler", url)
+	default:
+		cmd = exec.Command("xdg-open", url)
+	}
+	return cmd.Start()
+}